Divided Progression and Activity Log
Progression and Activity log are now in separate tabs and components (files) as well. No functions has been added or changed.
